chore: rename modal controller
