+++ resources/js/components/catalog-new/DisclosureModalController.vue
